+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [unreleased]
+
+### Added
+
+- Support cron job provisioning tasks[Provision all]
+
 ## [3.3.10-rciam] - 2021-12-14
 
 ### Fixed